backend/slippsserver/requirements.txt
Django==1.11.2
django-cors-headers==2.1.0
django-elasticsearch-dsl==0.1.0
django-filter==1.0.4
django-rest-swagger==2.1.2
django-webpack-loader==0.5.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
elasticsearch==6.1.1
elasticsearch-dsl==5.3.0
hashids==1.1.0
psycopg2==2.7.3.1
python-env==1.0.0
python-json-logger==0.1.7
python-logstash==0.4.6
djangorestframework==3.7.7
markdown==2.6.11
django-filter==1.1.0
